U-Net/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
Pillow>=9.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
albumentations>=1.2.0 
torchmetrics==1.7.3 
optuna>=3.0.0        
seaborn>=0.11.0